changing test sql commands to set build_id and id as bigint. Also change repoppush function inside of werker to use branch var

+++ util/storage/test-fixtures/schema.sql
@@ -10,14 +10,14 @@ CREATE TABLE build_summary (
 );
 
 CREATE TABLE build_output (
-  build_id INTEGER,
+  build_id BIGINT,
   output CHARACTER VARYING,
   id SERIAL PRIMARY KEY,
   FOREIGN KEY (build_id) REFERENCES build_summary (id) ON DELETE CASCADE
 );
 
 CREATE TABLE build_failure_reason (
-  build_id INTEGER,
+  build_id BIGINT,
   reasons jsonb,
   id SERIAL PRIMARY KEY,
   FOREIGN KEY (build_id) REFERENCES build_summary (id) ON DELETE CASCADE
